[IOTDB-5842] Sync: Delete BufferedPipeDataQueueTest (#9774)
Delete BufferedPipeDataQueueTest

This UT tests the BufferedPipeDataQueue, a buffered queue used by the original sync sender primarily as a producer-consumer model's component. The new version of the pipe system will replace the implementation of the sender with a three-stage flow model, which uses Disruptor as the cache queue. Therefore the UT for BufferedPipeDataQueue is not necessary.