[IOTDB-6209] Pipe: Solving the topological order of the progress index in the historical data collection phase (#11478)
* Problem:
When the pipe performs historical data collection, it currently sends sequential data first and then disorganized data, which is obviously wrong because the progress index of some of the disorganized data may be smaller than that of sequential files, which causes the pipe to record the wrong progress index as the progress information, resulting in some of the disorganized files not being sent.

* Solution:
After collecting all the historical data, solve for the topological order according to the progress index contained in the TsFile. Use the topological order of the file as the order in which the historical data is collected.