Pipe: skip pipe CN meta sync after successful synchronization when having no pipe (#11532)
In order to fix the issue that `sync` operation logs are still printed when no pipe exists in CN, like:

```text
2023-11-13 21:01:36,412 [ProcExecWorker-1] INFO  o.a.i.c.p.ProcedureExecutor:394 - pid=1, state=SUCCESS; org.apache.iotdb.confignode.procedure.impl.pipe.runtime.PipeMetaSyncProcedure finished in 55ms successfully.
```

a simple idea is to, in the next round of `sync` after a successful `sync` operation, check whether the current version of `PipeTaskInfo` is the same as the version synced before (empty `PipeTaskInfo`), and whether the current `PipeTaskInfo` is empty (indicating no need for a restart).

In terms of specific implementation:

1. Maintain the version of `PipeTaskInfo` synced by CN in the last round. In this context, the version is considered as the number of times the `PipeTaskInfo` write lock is acquired.
    - If a restart occurs, it is the version after `PipeHandleMetaChangeProcedure`; otherwise, it is the version after `PipeMetaSyncProcedure`.
    - Update in the lock release phase of these procedures to ensure that the recorded version at this point is free from concurrency issues.
    - Considering that `PipeHandleMetaChangeProcedure` may also be submitted by other logic, in order to avoid introducing new variables that may affect the serialization and deserialization logic in `PipeHandleMetaChangeProcedure`, we only update the version of `PipeTaskInfo` synced by `PipeMetaSyncProcedure` (partially affecting the liveness of **skipping sync** but not affecting correctness).
2. Maintain whether the last sync was successful in `PipeMetaSyncer`.
    - To check if the `sync` operation succeeded, it is also necessary to block and wait for the execution result of `PipeHandleMetaChangeProcedure` (refer to #11475).

Since querying `PipeTaskInfo` before each round of `sync` operations is done without locking, there may be other procedures modifying `PipeTaskInfo` at this time. After verification, it is confirmed that this only partially affects the liveness of **sync** but does not affect correctness.