Pipe: avoid useless pipe meta sync (stopped status) to DN to achieve high availability (#11641)
After #11279 https://github.com/apache/iotdb/pull/11279, when a pipe encounters a critical exception on DN, it stops all DR's pipe tasks related to that pipe and modifies the status of the relevant pipe to exception stopped on this DN. Subsequently, the CN collects this information through the heartbeat mechanism and pushes the changes to all DNs. This causes other DNs with running normal pipe tasks to be auto-restarted unnecessarily. This PR addresses this issue.

From now on, there will be a subtle change in the semantics of the CN-side pipe status. When the status is exception stopped, it indicates that the pipe tasks on some (not necessarily all) DRs (DNs) of this pipe encountered a critical exception.

----------

* chore: avoid unused pipe meta sync to DN

* fix: remove PipeHandleMetaChangeProcedure in start pipe

* chore: some logic optimizations

* chore: add comments for stopping pipe