Pipe: Utilize parallelStream for concurrent execution of create, start, stop, and drop pipe tasks to enhance performance (#11892)
Currently, when creating a pipe, if there is historical data in the cluster, the pipe will be automatically started. This involves **serially** extracting historical data in each data region. When dealing with large data, timeouts may occur, leading to pipe creation failure. To address this, we are considering parallelizing the above operations using `parallelStream` (inspired by @SteveYurongSu).