Pipe: distinguish between old and new process subtasks before and after stuck restart to avoid unexpected hashmap override and removal (#12400)
We noticed that when the pipe is stuck and restarted, the hashcode of new and old process subtasks is identical. This may lead to a situation where the key and value are inconsistent in the maintenance of ConcurrentHashMap `subtasks -> subtasks` by `PipeProcessorSubtaskWorker`, thereby causing unexpectedly removed subtasks that are NOT in the close state by `cleanupClosedSubtasksIfNecessary` in `org.apache.iotdb.db.pipe.task.subtask.processor.PipeProcessorSubtaskWorker`.